serverless/image_processor/requirements.txt
Pillow==10.4.0
boto3==1.34.162

# Type stubs for LSP completion
boto3-stubs[s3]==1.34.162
types-Pillow==10.2.0.20240520